:sparkles: 신고 DTO 수정

+++ general_project/general_project/Network/Entity/ReportModel.swift
@@ -97,7 +97,7 @@ struct MatchResponseData: Codable {
     let feedId: Int
     let authorId: Int
     let reporterId: Int
-    let similarity: Int
+    let similarity: Double
     let saved: Bool
     let message: String?
